Add __attribute__((nonnull)) in ft_io.h when it was pertinent [OK] Add every int types ato*_base                                 [OK] Rework every int types *toa + *toa_base                       [OK]

+++ srcs/ft_printf/cvrt/cvrt_s.c
@@ -6,7 +6,7 @@
 /*   By: jodufour <jodufour@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/09 04:38:14 by jodufour          #+#    #+#             */
-/*   Updated: 2021/11/15 23:58:57 by jodufour         ###   ########.fr       */
+/*   Updated: 2021/11/20 07:20:17 by jodufour         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int	padded_putnstr(
 	if (ctx->flags & (1 << 1)
 		&& padding('0', ctx->fwidth - ctx->prec))
 		return (MALLOC_ERR);
-	write(1, str, ctx->prec);
+	write(1, str, (size_t)ctx->prec);
 	if (ctx->flags & (1 << 0)
 		&& padding(' ', ctx->fwidth - ctx->prec))
 		return (MALLOC_ERR);
@@ -66,7 +66,7 @@ static int	get_char_ptr(t_ctx *const ctx, va_list va)
 	ctx->len += ctx->fwidth;
 	if (ctx->fwidth > ctx->prec)
 		return (padded_putnstr(str, ctx));
-	write(1, str, ctx->prec);
+	write(1, str, (size_t)ctx->prec);
 	return (SUCCESS);
 }
