+++ srcs/ft_printf/manage_text.c
@@ -6,7 +6,7 @@
 /*   By: jodufour <jodufour@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/09 06:37:06 by jodufour          #+#    #+#             */
-/*   Updated: 2021/11/16 00:00:41 by jodufour         ###   ########.fr       */
+/*   Updated: 2021/11/20 07:22:54 by jodufour         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,10 +20,10 @@ char	*manage_text(char const *format, t_ctx *const ctx)
 
 	next = int_strchr(format, '%');
 	if (next)
-		ctx->fwidth = next - format;
+		ctx->fwidth = (int)(next - format);
 	else
 		ctx->fwidth = int_strlen(format);
-	write(1, format, ctx->fwidth);
+	write(1, format, (size_t)ctx->fwidth);
 	ctx->len += ctx->fwidth;
 	format += ctx->fwidth;
 	return ((char *)format);
